+++ boot/authority-web-boot-starter/src/main/java/com/kewen/framework/boot/authority/service/impl/LoginServiceImpl.java
@@ -84,7 +84,7 @@ public class LoginServiceImpl implements LoginService {
         return loginResp;
     }
 
-    private UserDetail fetchUserDetail( Long userId,String userName) {
+    private UserDetail fetchUserDetail( Integer userId,String userName) {
         //查询机构
         SysUserDept sysUserDept = userDeptService.getOne(
                 new LambdaQueryWrapper<SysUserDept>()
